+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.5.0] - 2026-07-14
+
+### Added
+
+- Écran d'accueil retravaillé : logo en dégradé, panneau d'infos avec icônes, et nombre d'entrées affiché
+- README : section d'installation Linux pas à pas
+
 ## [1.4.2] - 2026-07-14
 
 ### Changed